polish

+++ framework/management/commands/deploy_themes.py
@@ -1,5 +1,6 @@
 #coding: utf-8
 
+import os.path
 import shutil
 
 from datetime import datetime
@@ -12,7 +13,8 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         target = settings.MEDIA_ROOT + "themes"
-        shutil.rmtree(target)
+        if os.path.isdir(target):
+            shutil.rmtree(target)
         shutil.copytree("./themes", target)
         print "done"
